requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
google-api-core==1.22.3
google-auth==1.22.0
google-cloud==0.34.0
google-cloud-core==1.4.2
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
numpy==1.19.2
pandas==1.1.2
pkg-resources==0.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.25.10
yarl==1.6.0
